Refactor code to avoid circular imports
