+++ cms/tests/templatetags.py
@@ -25,7 +25,7 @@ from cms.test_utils.fixtures.templatetags import TwoPagesFixture
 from cms.test_utils.testcases import CMSTestCase
 from cms.toolbar.toolbar import CMSToolbar
 from cms.utils import get_cms_setting, get_site_id
-from cms.utils.plugins import get_placeholders
+from cms.utils.placeholder import get_placeholders
 from sekizai.context import SekizaiContext
